deployment: fuck it, whatever...

+++ deployment/prepare_gameclient.py
@@ -77,16 +77,12 @@ with open(os.path.join(args.build_dir, args.data_file), 'wb') as f:
     f.write(data)
     f.close()
 
-remove_files = [
+ignored_files = [
     'OTPInternalRepository',
     'PiratesAIRepository',
     'PiratesUberRepository',
     'ServiceStart',
     'InventoryInit'
-]
-
-ignore_files = [
-    'CrewHUD'
 ]
 
 def cleanup_tree(directory):
@@ -98,18 +94,10 @@ def cleanup_tree(directory):
             if os.path.isdir(filepath):
                 recurse(filepath)
             elif os.path.isfile(filepath):
-                ignore_file = False
-                for filename in ignore_files:
-                    if filename in filepath:
-                        ignore_file = True
-
-                if ignore_file:
-                    continue
-
-                if filepath.endswith('AI.py') or filepath.endswith('UD.py'):
+                if (filepath.endswith('AI.py') or filepath.endswith('UD.py')) and 'CrewHUD' not in filepath:
                     os.remove(filepath)
 
-                for filename in remove_files:
+                for filename in ignored_files:
                     if filename in filepath:
                         os.remove(filepath)
                         break
